VPA: track list of managed pods on VPA object
Previously the VPA object didn't have any references to the pods it
was managing, only the count of those pods. This required us to do a lot
of looping over pods if we wanted to know any details about the pod
states of the pods that were being managed by a VPA.

This adds a ManagedPods map to the VPA object that contains references
to the cluster state for each of the pods the VPA is managing, which
lets us cheaply access them so we can do things like container cleanup.